% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanova.R
\name{permanovaModF}
\alias{permanovaModF}
\title{Calculates modified F statistic (Andersson 2017) and permutation p-value}
\usage{
permanovaModF(mat, metaData, groups = c(), permutations = 1000L)
}
\arguments{
\item{mat}{matrix of ordination data}

\item{metaData}{data_frame like object with columns with same name as \code{groups}}

\item{groups}{character vector containing colnames to test across}

\item{permutations}{integer number of permutations (including initial observation)}
}
\value{
A \code{data_frame} containing modified F and permutation p values
}
\description{
performs a one way ANOVA and calculates modified F statistic to account for unbalanced designs
with heterogeneous multivariate dispersion. Carries out a permutaiton test to calculate significance.
Based on the permanova function in Fathom Toolbox.
}
\details{
Authours: Chris Ward
}
