% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanova_pw.R
\name{permanovaModF_PW}
\alias{permanovaModF_PW}
\title{Calculates pairwise modified F statistic (Andersson 2017) and permutation p-value for
each comparison of factors within a group}
\usage{
permanovaModF_PW(mat, metaData, group = c(), permutations = 1000L)
}
\arguments{
\item{mat}{matrix of ordination data}

\item{metaData}{data_frame like object with one column with same name as \code{group}, must include
a \code{sample_id} column with sample id in the same order as \code{mat}}

\item{group}{character vector containing colname to test across}

\item{permutations}{integer number of permutations (including initial observation)}
}
\value{
A \code{data_frame} containing modified F and permutation p values
}
\description{
performs a pairwise one way ANOVA and calculates modified F statistic to account for unbalanced designs
with heterogeneous multivariate dispersion. Carries out a permutaiton test to calculate significance.
Based on the permanova function in Fathom Toolbox.
}
\details{
Authours: Chris Ward
}
